Rename configs file to config
